Deprecated MNTP now all configuration options are available on other property editors.
