+++ src/Umbraco.Core/PropertyEditors/DataEditorAttribute.cs
@@ -1,4 +1,4 @@
-namespace Umbraco.Cms.Core.PropertyEditors;
+﻿namespace Umbraco.Cms.Core.PropertyEditors;
 
 /// <summary>
 ///     Marks a class that represents a data editor.
