+++ src/Umbraco.Core/PropertyEditors/IDataEditor.cs
@@ -1,4 +1,4 @@
-using Umbraco.Cms.Core.Composing;
+﻿using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.Models;
 
 namespace Umbraco.Cms.Core.PropertyEditors;
