+++ src/Umbraco.Core/PropertyEditors/MediaPicker3Configuration.cs
@@ -1,4 +1,4 @@
-namespace Umbraco.Cms.Core.PropertyEditors;
+﻿namespace Umbraco.Cms.Core.PropertyEditors;
 
 /// <summary>
 ///     Represents the configuration for the media picker value editor.
