+++ src/Umbraco.Core/PropertyEditors/MultiNodePickerConfigurationEditor.cs
@@ -1,4 +1,4 @@
-// Copyright (c) Umbraco.
+﻿// Copyright (c) Umbraco.
 // See LICENSE for more details.
 
 using Umbraco.Cms.Core.IO;
@@ -8,6 +8,7 @@ namespace Umbraco.Cms.Core.PropertyEditors;
 /// <summary>
 ///     Represents the configuration for the multinode picker value editor.
 /// </summary>
+[Obsolete("The multi node tree picker is replaced by the dedicated document, media, element and member pickers. Scheduled for removal in Umbraco 22.")]
 public class MultiNodePickerConfigurationEditor : ConfigurationEditor<MultiNodePickerConfiguration>
 {
     /// <summary>
